Refactor model page and models subscription into a new store. Reset model store on models page unmount.
